refactor and adding MIT license for myself
why change(s) is important:
making it easier to use package, simplying maintenance for setup.py, and 
because i'm deleting/editing ~33K lines of code in repository.
