+++ README.md
@@ -1,5 +1,4 @@
-# Sastrawi Python
-===============
+# PySastrawi
 
 [![Build Status](https://travis-ci.org/har07/PySastrawi.svg?branch=master)](https://travis-ci.org/har07/PySastrawi)
 [![Coverage Status](https://coveralls.io/repos/github/har07/PySastrawi/badge.svg?branch=master)](https://coveralls.io/github/har07/PySastrawi?branch=master)
@@ -7,17 +6,17 @@
 
 ## English
 
-Sastrawi is a simple Python module which allows you to reduce inflected words in Indonesian Language (Bahasa Indonesia) to their base form ([stem](http://en.wikipedia.org/wiki/Stemming)). This is Python port of the original [Sastrawi](https://github.com/sastrawi/sastrawi) project written in PHP (credits goes to the original author and contributors of Sastrawi PHP).
+PySastrawi is a simple Python package which allows you to reduce inflected words in Indonesian Language (Bahasa Indonesia) to their base form ([stem](http://en.wikipedia.org/wiki/Stemming)). This is Python port of the original [Sastrawi](https://github.com/sastrawi/sastrawi) project written in PHP (credits goes to the original author and contributors of Sastrawi PHP).
 
 Installation
 ------------
-Sastrawi can be installed using [pip](https://docs.python.org/3.6/installing/index.html), by running `pip install PySastrawi` in terminal/command prompt.
+PySastrawi can be installed using [pip](https://docs.python.org/3.6/installing/index.html), by running `pip install PySastrawi` in terminal/command prompt.
 
 Usage
 ------
 
 ```python
-from Sastrawi.stemmer import *
+import Sastrawi
 
 stemmer = Stemmer()
 
@@ -37,16 +36,10 @@ for sentence in sentences:
 # meniru-nirukannya
 
 # more about stemming process
-res = Context("pemerintahannya", stemmer.stemwords, stemmer.visitor_provider)
+res = stemmer.context('kemerdekaannya')
 
 print(res.result)
-# perintah
-
-for i in res.removals:
-    print(i)
-# Removal(subject='pemerintahannya', result='pemerintahan', removedPart='nya', affixType='PP')
-# Removal(subject='pemerintahan', result='pemerintah', removedPart='an', affixType='DS')
-# Removal(subject='pemerintah', result='perintah', removedPart='pemerintah', affixType='DP')
+# ('merdeka', [('nya', 'PP'), ('an', 'DS'), ('ke', 'DP')])
 
 ```
 
@@ -72,18 +65,18 @@ More information
 
 ## Indonesia
 
-Sastrawi adalah module Python sederhana yang memungkinkan anda untuk mengubah kata berimbuhan dalam Bahasa Indonesia ke bentuk kata dasar ([stem](http://en.wikipedia.org/wiki/Stemming)). Proyek ini adalah port dari proyek [Sastrawi](https://github.com/sastrawi/sastrawi) yang ditulis dalam PHP (kredit untuk pencipta dan semua kontributor Sastrawi PHP)
+PySastrawi adalah package Python sederhana yang memungkinkan anda untuk mengubah kata berimbuhan dalam Bahasa Indonesia ke bentuk kata dasar ([stem](http://en.wikipedia.org/wiki/Stemming)). Proyek ini adalah port dari proyek [Sastrawi](https://github.com/sastrawi/sastrawi) yang ditulis dalam PHP (kredit untuk pencipta dan semua kontributor Sastrawi PHP)
 
 Cara Install
 -------------
 
-Sastrawi dapat di-*install* menggunakan [pip](https://docs.python.org/3.6/installing/index.html), dengan menjalankan perintah `pip install PySastrawi` di terminal/command prompt.
+PySastrawi dapat di-*install* menggunakan [pip](https://docs.python.org/3.6/installing/index.html), dengan menjalankan perintah `pip install PySastrawi` di terminal/command prompt.
 
 Penggunaan
 -----------
 
 ```python
-from Sastrawi.stemmer import *
+import Sastrawi
 
 stemmer = Stemmer()
 
@@ -103,17 +96,10 @@ for sentence in sentences:
 # meniru-nirukannya
 
 # lebih dalam tentang proses stemming
-res = Context("pemerintahannya", stemmer.stemwords, stemmer.visitor_provider)
+res = stemmer.context('kemerdekaannya')
 
 print(res.result)
-# perintah
-
-for i in res.removals:
-    print(i)
-# Removal(subject='pemerintahannya', result='pemerintahan', removedPart='nya', affixType='PP')
-# Removal(subject='pemerintahan', result='pemerintah', removedPart='an', affixType='DS')
-# Removal(subject='pemerintah', result='perintah', removedPart='pemerintah', affixType='DP')
-
+# ('merdeka', [('nya', 'PP'), ('an', 'DS'), ('ke', 'DP')])
 ```
 
 Demo
@@ -125,7 +111,7 @@ Repository : https://github.com/har07/pystastrawi-demo
 Lisensi
 --------
 
-Lisensi Sastrawi adalah MIT License (MIT).
+Lisensi PySastrawi adalah MIT License (MIT).
 
 Project ini mengandung kamus kata dasar yang berasal dari Kateglo dengan lisensi
 [CC-BY-NC-SA 3.0](http://creativecommons.org/licenses/by-nc-sa/3.0/).
@@ -133,4 +119,4 @@ Project ini mengandung kamus kata dasar yang berasal dari Kateglo dengan lisensi
 Informasi Lebih Lanjut
 ----------------------
 
-- [Sastrawi PHP Repository page](https://github.com/sastrawi/sastrawi)
+- [Halaman Repository Sastrawi PHP](https://github.com/sastrawi/sastrawi)
