+++ src/Sastrawi/__init__.py
@@ -1,0 +1,6 @@
+"""
+Sastrawi is a simple Python package which allows you to reduce inflected 
+words in Indonesian Language (Bahasa Indonesia) to their base form.
+"""
+
+from Sastrawi.stemming import Stemmer
