+++ src/Sastrawi/data/rootwords.txt
